+++ tests/test_extensions/test_ext_viewcode.py
@@ -1,23 +1,19 @@
 """Test sphinx.ext.viewcode extension."""
 
-from __future__ import annotations
-
 import re
-import shutil
-from typing import TYPE_CHECKING
 
 import pytest
 
-if TYPE_CHECKING:
-    from sphinx.testing.util import SphinxTestApp
 
+@pytest.mark.sphinx(testroot='ext-viewcode')
+def test_viewcode(app, status, warning):
+    app.builder.build_all()
 
-def check_viewcode_output(app: SphinxTestApp) -> str:
-    warnings = re.sub(r'\\+', '/', app.warning.getvalue())
+    warnings = re.sub(r'\\+', '/', warning.getvalue())
     assert re.findall(
-        r"index.rst:\d+: WARNING: Object named 'func1' not found in include "
+        r"index.rst:\d+: WARNING: Object named 'func1' not found in include " +
         r"file .*/spam/__init__.py'",
-        warnings,
+        warnings
     )
 
     result = (app.outdir / 'index.html').read_text(encoding='utf8')
@@ -34,56 +30,20 @@ def check_viewcode_output(app: SphinxTestApp) -> str:
     assert result.count('this is the class attribute class_attr') == 2
 
     result = (app.outdir / '_modules/spam/mod1.html').read_text(encoding='utf8')
-    # filter pygments classes
-    result = re.sub('<span class="[^"]{,2}">', '<span>', result)
-    assert (
-        '<div class="viewcode-block" id="Class1">\n'
-        '<a class="viewcode-back" href="../../index.html#spam.Class1">[docs]</a>\n'
-    ) in result
-    assert '<span>@decorator</span>\n' in result
-    assert '<span>class</span> <span>Class1</span><span>:</span>\n' in result
-    assert '<span>    </span><span>&quot;&quot;&quot;</span>\n' in result
-    assert '<span>    this is Class1</span>\n' in result
-    assert '<span>    &quot;&quot;&quot;</span>\n' in result
-
-    return result
-
-
-@pytest.mark.sphinx(
-    'html',
-    testroot='ext-viewcode',
-    freshenv=True,
-    confoverrides={'viewcode_line_numbers': True},
-)
-@pytest.mark.usefixtures('rollback_sysmodules')
-def test_viewcode_linenos(app):
-    shutil.rmtree(app.outdir / '_modules', ignore_errors=True)
-    app.build(force_all=True)
-
-    result = check_viewcode_output(app)
-    assert '<span class="linenos"> 1</span>' in result
-
-
-@pytest.mark.sphinx(
-    'html',
-    testroot='ext-viewcode',
-    freshenv=True,
-    confoverrides={'viewcode_line_numbers': False},
-)
-@pytest.mark.usefixtures('rollback_sysmodules')
-def test_viewcode(app):
-    shutil.rmtree(app.outdir / '_modules', ignore_errors=True)
-    app.build(force_all=True)
-
-    result = check_viewcode_output(app)
-    assert 'class="linenos">' not in result
+    result = re.sub('<span class=".*?">', '<span>', result)  # filter pygments classes
+    assert ('<div class="viewcode-block" id="Class1"><a class="viewcode-back" '
+            'href="../../index.html#spam.Class1">[docs]</a>'
+            '<span>@decorator</span>\n'
+            '<span>class</span> <span>Class1</span>'
+            '<span>(</span><span>object</span><span>):</span>\n'
+            '    <span>&quot;&quot;&quot;</span>\n'
+            '<span>    this is Class1</span>\n'
+            '<span>    &quot;&quot;&quot;</span></div>\n') in result
 
 
 @pytest.mark.sphinx('epub', testroot='ext-viewcode')
-@pytest.mark.usefixtures('rollback_sysmodules')
-def test_viewcode_epub_default(app):
-    shutil.rmtree(app.outdir)
-    app.build(force_all=True)
+def test_viewcode_epub_default(app, status, warning):
+    app.builder.build_all()
 
     assert not (app.outdir / '_modules/spam/mod1.xhtml').exists()
 
@@ -91,14 +51,10 @@ def test_viewcode_epub_default(app):
     assert result.count('href="_modules/spam/mod1.xhtml#func1"') == 0
 
 
-@pytest.mark.sphinx(
-    'epub',
-    testroot='ext-viewcode',
-    confoverrides={'viewcode_enable_epub': True},
-)
-@pytest.mark.usefixtures('rollback_sysmodules')
-def test_viewcode_epub_enabled(app):
-    app.build(force_all=True)
+@pytest.mark.sphinx('epub', testroot='ext-viewcode',
+                    confoverrides={'viewcode_enable_epub': True})
+def test_viewcode_epub_enabled(app, status, warning):
+    app.builder.build_all()
 
     assert (app.outdir / '_modules/spam/mod1.xhtml').exists()
 
@@ -106,23 +62,23 @@ def test_viewcode_epub_enabled(app):
     assert result.count('href="_modules/spam/mod1.xhtml#func1"') == 2
 
 
-@pytest.mark.sphinx('html', testroot='ext-viewcode', tags=['test_linkcode'])
-def test_linkcode(app):
-    app.build(filenames=[app.srcdir / 'objects.rst'])
+@pytest.mark.sphinx(testroot='ext-viewcode', tags=['test_linkcode'])
+def test_linkcode(app, status, warning):
+    app.builder.build(['objects'])
 
     stuff = (app.outdir / 'objects.html').read_text(encoding='utf8')
 
-    assert 'https://foobar/source/foolib.py' in stuff
-    assert 'https://foobar/js/' in stuff
-    assert 'https://foobar/c/' in stuff
-    assert 'https://foobar/cpp/' in stuff
+    assert 'http://foobar/source/foolib.py' in stuff
+    assert 'http://foobar/js/' in stuff
+    assert 'http://foobar/c/' in stuff
+    assert 'http://foobar/cpp/' in stuff
 
 
-@pytest.mark.sphinx('html', testroot='ext-viewcode-find', freshenv=True)
-def test_local_source_files(app):
+@pytest.mark.sphinx(testroot='ext-viewcode-find')
+def test_local_source_files(app, status, warning):
     def find_source(app, modname):
         if modname == 'not_a_package':
-            source = app.srcdir / 'not_a_package/__init__.py'
+            source = (app.srcdir / 'not_a_package/__init__.py').read_text(encoding='utf8')
             tags = {
                 'func1': ('def', 1, 1),
                 'Class1': ('class', 1, 1),
@@ -130,7 +86,7 @@ def test_local_source_files(app):
                 'not_a_package.submodule.Class1': ('class', 1, 1),
             }
         else:
-            source = app.srcdir / 'not_a_package/submodule.py'
+            source = (app.srcdir / 'not_a_package/submodule.py').read_text(encoding='utf8')
             tags = {
                 'not_a_package.submodule.func1': ('def', 11, 15),
                 'Class1': ('class', 19, 22),
@@ -138,26 +94,33 @@ def test_local_source_files(app):
                 'Class3': ('class', 25, 30),
                 'not_a_package.submodule.Class3.class_attr': ('other', 29, 29),
             }
-        return source.read_text(encoding='utf8'), tags
+        return (source, tags)
 
     app.connect('viewcode-find-source', find_source)
-    app.build(force_all=True)
+    app.builder.build_all()
 
-    warnings = re.sub(r'\\+', '/', app.warning.getvalue())
+    warnings = re.sub(r'\\+', '/', warning.getvalue())
     assert re.findall(
-        r"index.rst:\d+: WARNING: Object named 'func1' not found in include "
+        r"index.rst:\d+: WARNING: Object named 'func1' not found in include " +
         r"file .*/not_a_package/__init__.py'",
-        warnings,
+        warnings
     )
 
     result = (app.outdir / 'index.html').read_text(encoding='utf8')
-    for needle in (
-        'href="_modules/not_a_package.html#func1"',
-        'href="_modules/not_a_package.html#not_a_package.submodule.func1"',
-        'href="_modules/not_a_package/submodule.html#Class1"',
-        'href="_modules/not_a_package/submodule.html#Class3"',
-        'href="_modules/not_a_package/submodule.html#not_a_package.submodule.Class1"',
-        'href="_modules/not_a_package/submodule.html#not_a_package.submodule.Class3.class_attr"',
-        'This is the class attribute class_attr',
-    ):
-        assert result.count(needle) == 1
+    assert result.count('href="_modules/not_a_package.html#func1"') == 1
+    assert result.count('href="_modules/not_a_package.html#not_a_package.submodule.func1"') == 1
+    assert result.count('href="_modules/not_a_package/submodule.html#Class1"') == 1
+    assert result.count('href="_modules/not_a_package/submodule.html#Class3"') == 1
+    assert result.count('href="_modules/not_a_package/submodule.html#not_a_package.submodule.Class1"') == 1
+
+    assert result.count('href="_modules/not_a_package/submodule.html#not_a_package.submodule.Class3.class_attr"') == 1
+    assert result.count('This is the class attribute class_attr') == 1
+
+
+@pytest.mark.sphinx(testroot='ext-viewcode-find-package')
+def test_local_source_files(app, status, warning):
+    app.builder.build_all()
+    result = (app.outdir / 'index.html').read_text(encoding='utf8')
+    assert result.count('href="_modules/main_package/subpackage/_subpackage2/submodule.html#func1"') == 1
+    assert result.count('href="_modules/main_package/subpackage/_subpackage2/submodule.html#Class1"') == 1
+    assert result.count('href="_modules/main_package/subpackage/_subpackage2/submodule.html#Class3"') == 1
